update AISKU perf tests
