+++ AISKU/Tests/Perf/src/aiskuperftests.ts
@@ -1,4 +1,6 @@
 import { AISKUPerf } from "./AISKUPerf.Tests";
+//Do not remove this import, it will be used in browser.
+import { Snippet } from "../../../src/Initialization";
 
 export function runTests() {
     new AISKUPerf().registerTests();
