+++ AISKU/package.json
@@ -20,7 +20,7 @@
         "build:snippet": "grunt snippetvnext",
         "rebuild": "npm run build",
         "test": "grunt aiskuunittests",
-        "perftest": "grunt aiskuperftests",
+        "perftest": "grunt aiskuperf",
         "lint": "tslint -p tsconfig.json",
         "dtsgen": "api-extractor run --local && node ../scripts/dtsgen.js 'Microsoft.ApplicationInsights'",
         "sri": "node ../tools/subResourceIntegrity/generateIntegrityFile.js",
